Proper mayastor debug package, right package versions, cleanup images.
There are two flavours of mayastor nix package: mayastor and
mayastor-dev. And they don't get accidentaly confused by the build
system as was happening before with "release" parameter.

Overhaul of nix code for creating docker images. No code duplication
and intuitive target names. All image targets end with "-image" suffix.
The flavour of the package that the image is based on (if any) preceeds
the suffix (i.e. dev or adhoc).

We are able to deduce the accurate version number of mayastor package
in effortless way. It uses a git command to lookup a tag for the commit
and if tag does not exist, it uses short commit hash. So it never gets
stale.