Add support for block volumes. CAS-14
To support block volumes.
  at node stage volume
    - attach to the nexus
  at node publish volume
    - create a block special file
    - bind mount the device path to the block special file
  at node unpublish volume
    - unmount the block special file
    - delete the block special file
  at node unstage volume
    - detach from the nexus

The CSI spec does not provide enough context at node unstage volume,
to determine whether the volume was a block volume or filesystem volume.
To handle this the mount point for filesystem volumes is move to the
subdirectory fs_mnt of the staging directory supplied CO.
At unstage we use the presence of this directory to differentiate,
between filesystem and block volumes.